// ui/lib/api/schemas/textSelectionShape.ts
/**
 * Generated by orval v8.19.0 🍺
 * Do not edit manually.
 */
import type { BlobRef } from './blobRef';
import type { Transform } from './transform';

export type TextSelectionShape = {
  kind: 'rectangle';
  transform: Transform;
} | {
  kind: 'brush';
  mask: BlobRef;
} | {
  kind: 'polygon';
  points: number[][];
};